refactor: import/export to entry points
