+++ frontend/src/components/organisms/ServerTable.tsx
@@ -28,7 +28,12 @@ export const ServerTable: React.FC = () => {
         </thead>
         <tbody>
           {cells.map((cell) => (
-            <CommonTable key={cell.id} editText={cell.usage} id={cell.id} location="server">
+            <CommonTable
+              key={cell.id}
+              editText={cell.usage}
+              id={cell.id}
+              location="server"
+            >
               <TableBodyMole>{cell.hostname}</TableBodyMole>
               <TableBodyMole>{cell.ip}</TableBodyMole>
               <TableBodyMole>{cell.username}</TableBodyMole>
