+++ frontend/src/components/organisms/index.ts
@@ -4,3 +4,4 @@ export { HomeTable } from './HomeTable'
 export { ServerTable } from './ServerTable'
 export { PwCell } from './PwCell'
 export { SideBar } from './SideBar'
+export { CommonForm } from './CommonForm'
