+++ frontend/src/components/templates/index.ts
@@ -1,2 +1,6 @@
 export { HomeTemp } from './HomeTemp'
+export { HomeEdit } from './HomeEdit'
+export { HomeNew } from './HomeNew'
+export { ServerNew } from './ServerNew'
+export { ServerEdit } from './ServerEdit'
 export { ServerTemp } from './ServerTemp'
