+++ frontend/src/types/index.ts
@@ -1,4 +1,4 @@
 export type { StatusState, SetStatusState, SetUserIdState } from 'types/context'
-export type { FormInput, Message } from 'types/form'
+export type { FormInput, Message, ServerFormInput } from 'types/form'
 export type { GetPws, GetPw } from 'types/pw'
 export type { GetUserId } from 'types/user'
